+++ app/src/main/java/pl/lambada/songsync/data/remote/UpdateService.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import kotlinx.coroutines.flow.flow
 import pl.lambada.songsync.data.remote.github.GithubAPI
 import pl.lambada.songsync.domain.model.Release
-import pl.lambada.songsync.ui.screens.about.components.UpdateState
 import pl.lambada.songsync.util.ext.getVersion
 
 class UpdateService {
@@ -43,4 +42,15 @@ class UpdateService {
 
         return latestVersion > currentVersion
     }
+}
+
+/**
+ * Defines the state of the update check.
+ */
+sealed interface UpdateState {
+    data object Idle : UpdateState
+    data object Checking : UpdateState
+    data object UpToDate : UpdateState
+    data class UpdateAvailable(val release: Release) : UpdateState
+    data class Error(val reason: Throwable) : UpdateState
 }
